requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
mysqlclient==2.1.1
